Move rpcMethods back to jbrowse-web
