+++ packages/jbrowse-web/src/jbrowseModel.js
@@ -2,7 +2,7 @@ import { ConfigurationSchema } from '@gmod/jbrowse-core/configuration'
 import PluginManager from '@gmod/jbrowse-core/PluginManager'
 import RpcManager from '@gmod/jbrowse-core/rpc/RpcManager'
 import { getSnapshot, resolveIdentifier, types } from 'mobx-state-tree'
-import * as rpcFuncs from '@gmod/jbrowse-core/rpc/rpcMethods'
+import * as rpcFuncs from './rpcMethods'
 import assemblyManager from './assemblyManager'
 import AssemblyConfigSchemasFactory from './assemblyConfigSchemas'
 import corePlugins from './corePlugins'
