+++ packages/jbrowse-web/src/rpcMethods.js
@@ -1,9 +1,12 @@
-import { checkAbortSignal } from '../util'
-import { freeAdapterResources, getAdapter } from '../util/dataAdapterCache'
+import { checkAbortSignal } from '@gmod/jbrowse-core/util'
+import {
+  freeAdapterResources,
+  getAdapter,
+} from '@gmod/jbrowse-core/util/dataAdapterCache'
 import {
   deserializeAbortSignal,
   isRemoteAbortSignal,
-} from './remoteAbortSignals'
+} from '@gmod/jbrowse-core/rpc/remoteAbortSignals'
 
 export async function getGlobalStats(
   pluginManager,
