+++ packages/jbrowse-web/src/rpcMethods.test.js
